MainActivity削除と細かいところの修正
