+++ Android/Panorama360/app/src/main/AndroidManifest.xml
@@ -13,19 +13,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Panorama360">
         <activity
-            android:name=".MainActivity2"
-            android:exported="true"
-            android:label="@string/title_activity_main2"
-            android:theme="@style/Theme.Panorama360">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity
             android:name=".MainActivity"
-            android:exported="true">
+            android:exported="true"
+            android:theme="@style/Theme.Panorama360">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
